[Verda] Rename the `datacrunch` backend to `verda` #3344
