+++ src/dstack/_internal/core/backends/datacrunch/__init__.py
@@ -1,0 +1,1 @@
+# DataCrunch backend for backward compatibility
