+++ src/tests/_internal/core/backends/datacrunch/test_configurator.py
@@ -1,21 +1,22 @@
 from unittest.mock import patch
 
-from dstack._internal.core.backends.datacrunch.configurator import (
-    DataCrunchConfigurator,
+from dstack._internal.core.backends.verda.configurator import (
+    VerdaConfigurator,
 )
-from dstack._internal.core.backends.datacrunch.models import (
-    DataCrunchBackendConfigWithCreds,
-    DataCrunchCreds,
+from dstack._internal.core.backends.verda.models import (
+    VerdaBackendConfigWithCreds,
+    VerdaCreds,
 )
 
 
-class TestDataCrunchConfigurator:
+class TestVerdaConfigurator:
     def test_validate_config_valid(self):
-        config = DataCrunchBackendConfigWithCreds(
-            creds=DataCrunchCreds(client_id="valid", client_secret="valid"),
+        config = VerdaBackendConfigWithCreds(
+            type="verda",
+            creds=VerdaCreds(client_id="valid", client_secret="valid"),
             regions=["FIN-01"],
         )
         with patch(
-            "dstack._internal.core.backends.datacrunch.configurator.DataCrunchConfigurator._validate_creds"
+            "dstack._internal.core.backends.verda.configurator.VerdaConfigurator._validate_creds"
         ):
-            DataCrunchConfigurator().validate_config(config, default_creds_enabled=True)
+            VerdaConfigurator().validate_config(config, default_creds_enabled=True)
